Catch Exception when attempting lock
When Hazelcast attempts to acquire a lock, it will block
until the lock is available. If another cluster member
has the lock and that member times out of the cluster,
an exception is thrown by `tryLock`.

This fix will catch any exception (not just `InterruptedException`)
to force a retry of lock acquisition.

See https://github.com/spring-cloud/spring-cloud-cluster/issues/24